cbm/frontend/cbm/requirements.txt
appdirs==1.4.0
Django==1.10.5
django-bootstrap3==8.1.0
django-controlcenter==0.2.2
django-pkgconf==0.2.0
django-tables2==1.3.0
gunicorn==19.6.0
mysqlclient==1.3.9
packaging==16.8
pyparsing==2.1.10
six==1.10.0
